// frontend/src/adapter/api/__generated/api.ts
export interface Item {
    /**
     * 
     * @type {string}
     * @memberof Item
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof Item
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof Item
     */
    'description'?: string;
}
// ...
export interface PostShoppingListsRequest {
    /**
     * 
     * @type {string}
     * @memberof PostShoppingListsRequest
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof PostShoppingListsRequest
     */
    'description'?: string;
    /**
     * 
     * @type {Array<PostShoppingListsRequestItemsInner>}
     * @memberof PostShoppingListsRequest
     */
    'items'?: Array<PostShoppingListsRequestItemsInner>;
}
/**
 * 
 * @export
 * @interface PostShoppingListsRequestItemsInner
 */
export interface PostShoppingListsRequestItemsInner {
    /**
     * 
     * @type {string}
     * @memberof PostShoppingListsRequestItemsInner
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof PostShoppingListsRequestItemsInner
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof PostShoppingListsRequestItemsInner
     */
    'description'?: string;
}
/**
 * 
 * @export
 * @interface PutItemsIdRequest
 */
export interface PutItemsIdRequest {
    /**
     * 
     * @type {string}
     * @memberof PutItemsIdRequest
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof PutItemsIdRequest
     */
    'description'?: string;
}
/**
 * 
 * @export
 * @interface PutShoppingListsIdRequest
 */
export interface PutShoppingListsIdRequest {
    /**
     * 
     * @type {string}
     * @memberof PutShoppingListsIdRequest
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof PutShoppingListsIdRequest
     */
    'description'?: string;
    /**
     * 
     * @type {Array<PutShoppingListsIdRequestItemsInner>}
     * @memberof PutShoppingListsIdRequest
     */
    'items'?: Array<PutShoppingListsIdRequestItemsInner>;
}
/**
 * 
 * @export
 * @interface PutShoppingListsIdRequestItemsInner
 */
export interface PutShoppingListsIdRequestItemsInner {
    /**
     * 
     * @type {string}
     * @memberof PutShoppingListsIdRequestItemsInner
     */
    'id': string;
    /**
     * 
     * @type {number}
     * @memberof PutShoppingListsIdRequestItemsInner
     */
    'quantity'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof PutShoppingListsIdRequestItemsInner
     */
    'isPurchased'?: boolean;
}
/**
 * 
 * @export
 * @interface PutShoppingListsShoppingListIdFavoritesRequest
 */
export interface PutShoppingListsShoppingListIdFavoritesRequest {
    /**
     * 
     * @type {boolean}
     * @memberof PutShoppingListsShoppingListIdFavoritesRequest
     */
    'isFavorite': boolean;
}
/**
 * 
 * @export
 * @interface PutShoppingListsShoppingListIdItemsItemIdRequest
 */
export interface PutShoppingListsShoppingListIdItemsItemIdRequest {
    /**
     * 
     * @type {number}
     * @memberof PutShoppingListsShoppingListIdItemsItemIdRequest
     */
    'quantity': number;
    /**
     * 
     * @type {boolean}
     * @memberof PutShoppingListsShoppingListIdItemsItemIdRequest
     */
    'isPurchased'?: boolean;
}
/**
 * 
 * @export
 * @interface ShoppingList
 */
export interface ShoppingList {
    /**
     * 
     * @type {string}
     * @memberof ShoppingList
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof ShoppingList
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof ShoppingList
     */
    'description'?: string;
    /**
     * 
     * @type {string}
     * @memberof ShoppingList
     */
    'createdAt': string;
    /**
     * 
     * @type {boolean}
     * @memberof ShoppingList
     */
    'isFavorite': boolean;
    /**
     * 
     * @type {Array<ShoppingListShoppingListItemsInner>}
     * @memberof ShoppingList
     */
    'shoppingListItems'?: Array<ShoppingListShoppingListItemsInner>;
}
// ...
export interface ShoppingListShoppingListItemsInner {
    /**
     * 
     * @type {boolean}
     * @memberof ShoppingListShoppingListItemsInner
     */
    'isPurchased'?: boolean;
    /**
     * 
     * @type {number}
     * @memberof ShoppingListShoppingListItemsInner
     */
    'quantity'?: number;
    /**
     * 
     * @type {Item}
     * @memberof ShoppingListShoppingListItemsInner
     */
    'item'?: Item;
}